+++ osx/SDLMain.m
@@ -311,9 +311,9 @@ static void CustomApplicationMain (int argc, char **argv)
 
 - (NSString *)stringByReplacingRange:(NSRange)aRange with:(NSString *)aString
 {
-    unsigned int bufferSize;
-    unsigned int selfLen = [self length];
-    unsigned int aStringLen = [aString length];
+    unsigned long int bufferSize;
+    unsigned long int selfLen = [self length];
+    unsigned long int aStringLen = [aString length];
     unichar *buffer;
     NSRange localRange;
     NSString *result;
